+++ project/Build.scala
@@ -1,45 +1,80 @@
+/*
+ * Copyright 2013 Twitter inc.
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
 package storehaus
 
 import sbt._
 import Keys._
-import sbtgitflow.ReleasePlugin._
 import spray.boilerplate.BoilerplatePlugin.Boilerplate
 import com.typesafe.tools.mima.plugin.MimaPlugin.mimaDefaultSettings
 import com.typesafe.tools.mima.plugin.MimaKeys.previousArtifact
 
 object StorehausBuild extends Build {
-  val extraSettings =
-    Project.defaultSettings ++ releaseSettings ++ Boilerplate.settings ++ mimaDefaultSettings
-  val sharedSettings =  extraSettings ++ Seq(
-    organization := "com.twitter",
-    crossScalaVersions := Seq("2.9.2", "2.10.0"),
-    libraryDependencies ++= Seq(
-      "org.scalacheck" %% "scalacheck" % "1.10.0" % "test" withSources(),
-      "org.scala-tools.testing" %% "specs" % "1.6.9" % "test" withSources()
-    ),
+  def withCross(dep: ModuleID) =
+    dep cross CrossVersion.binaryMapped {
+      case "2.9.3" => "2.9.2" // TODO: hack because twitter hasn't built things against 2.9.3
+      case version if version startsWith "2.10" => "2.10" // TODO: hack because sbt is broken
+      case x => x
+    }
 
+  def specs2Import(scalaVersion: String) = scalaVersion match {
+      case version if version startsWith "2.9" => "org.specs2" %% "specs2" % "1.12.4.1" % "test"
+      case version if version startsWith "2.10" => "org.specs2" %% "specs2" % "1.13" % "test"
+  }
+  val extraSettings =
+    Project.defaultSettings ++ Boilerplate.settings ++ mimaDefaultSettings
+
+  def ciSettings: Seq[Project.Setting[_]] =
+    if (sys.env.getOrElse("TRAVIS", "false").toBoolean) Seq(
+      ivyLoggingLevel := UpdateLogging.Quiet,
+      logLevel in Global := Level.Warn,
+      logLevel in Compile := Level.Warn,
+      logLevel in Test := Level.Info
+    ) else Seq.empty[Project.Setting[_]]
+
+  val testCleanup = Seq(
+    testOptions in Test += Tests.Cleanup { loader =>
+      val c = loader.loadClass("com.twitter.storehaus.testing.Cleanup$")
+      c.getMethod("cleanup").invoke(c.getField("MODULE$").get(c))
+    }
+  )
+
+  val sharedSettings = extraSettings ++ ciSettings ++ Seq(
+    organization := "com.twitter",
+    scalaVersion := "2.9.3",
+    crossScalaVersions := Seq("2.9.3", "2.10.0"),
+    javacOptions ++= Seq("-source", "1.6", "-target", "1.6"),
+    javacOptions in doc := Seq("-source", "1.6"),
+    libraryDependencies <+= scalaVersion(specs2Import(_)),
     resolvers ++= Seq(
       Opts.resolver.sonatypeSnapshots,
       Opts.resolver.sonatypeReleases,
       "Twitter Maven" at "http://maven.twttr.com"
     ),
-
     parallelExecution in Test := true,
-
     scalacOptions ++= Seq(Opts.compile.unchecked, Opts.compile.deprecation),
 
     // Publishing options:
     publishMavenStyle := true,
-
     publishArtifact in Test := false,
-
     pomIncludeRepository := { x => false },
-
     publishTo <<= version { v =>
       Some(if (v.trim.toUpperCase.endsWith("SNAPSHOT")) Opts.resolver.sonatypeSnapshots
            else Opts.resolver.sonatypeStaging)
     },
-
     pomExtra := (
       <url>https://github.com/twitter/storehaus</url>
       <licenses>
@@ -71,16 +106,17 @@ object StorehausBuild extends Build {
   /**
     * This returns the youngest jar we released that is compatible with
     * the current.
-    *
-    * TODO: Remove mysql, redis and cache after their release.
     */
+  val unreleasedModules = Set[String]()
+
   def youngestForwardCompatible(subProj: String) =
     Some(subProj)
-      .filterNot(Set("mysql", "redis", "cache", "zookeeper").contains(_))
-      .map { s => "com.twitter" % ("storehaus-" + s + "_2.9.2") % "0.2.0" }
+      .filterNot(unreleasedModules.contains(_))
+      .map { s => "com.twitter" % ("storehaus-" + s + "_2.9.3") % "0.7.0" }
 
-  val algebirdVersion = "0.1.12"
-  val bijectionVersion = "0.3.0"
+  val algebirdVersion = "0.3.0"
+  val bijectionVersion = "0.5.4"
+  val utilVersion = "6.8.2"
 
   lazy val storehaus = Project(
     id = "storehaus",
@@ -96,75 +132,57 @@ object StorehausBuild extends Build {
     storehausAlgebra,
     storehausMemcache,
     storehausMySQL,
-    storehausRedis
+    storehausRedis,
+    storehausHBase,
+    storehausDynamoDB,
+    storehausTesting
   )
 
-  lazy val storehausCache = Project(
-    id = "storehaus-cache",
-    base = file("storehaus-cache"),
-    settings = sharedSettings
-  ).settings(
-    name := "storehaus-cache",
-    previousArtifact := youngestForwardCompatible("cache")
-  )
+  def module(name: String) = {
+    val id = "storehaus-%s".format(name)
+    Project(id = id, base = file(id), settings = sharedSettings ++ testCleanup ++ Seq(
+      Keys.name := id,
+      previousArtifact := youngestForwardCompatible(name))
+    ).dependsOn(storehausTesting % "test->test")
+  }
 
-  lazy val storehausCore = Project(
-    id = "storehaus-core",
-    base = file("storehaus-core"),
-    settings = sharedSettings
-  ).settings(
-    name := "storehaus-core",
-    previousArtifact := youngestForwardCompatible("core"),
-    libraryDependencies += "com.twitter" %% "util-core" % "6.3.0",
-    libraryDependencies += "com.twitter" %% "bijection-core" % bijectionVersion
-  ).dependsOn(storehausCache)
+  lazy val storehausCache = module("cache")
 
-  lazy val storehausAlgebra = Project(
-    id = "storehaus-algebra",
-    base = file("storehaus-algebra"),
-    settings = sharedSettings
-  ).settings(
-    name := "storehaus-algebra",
-    previousArtifact := youngestForwardCompatible("algebra"),
+  lazy val storehausCore = module("core").settings(
+    libraryDependencies ++= Seq(
+      withCross("com.twitter" %% "util-core" % utilVersion),
+      "com.twitter" %% "bijection-core" % bijectionVersion,
+      "com.twitter" %% "bijection-util" % bijectionVersion
+    )
+  ).dependsOn(storehausCache %  "test->test;compile->compile")
+
+  lazy val storehausAlgebra = module("algebra").settings(
     libraryDependencies += "com.twitter" %% "algebird-core" % algebirdVersion,
     libraryDependencies += "com.twitter" %% "algebird-util" % algebirdVersion,
-    libraryDependencies += "com.twitter" %% "bijection-algebird" % bijectionVersion
+    libraryDependencies += "com.twitter" %% "algebird-bijection" % algebirdVersion
   ).dependsOn(storehausCore % "test->test;compile->compile")
 
-  lazy val storehausMemcache = Project(
-    id = "storehaus-memcache",
-    base = file("storehaus-memcache"),
-    settings = sharedSettings
-  ).settings(
-    name := "storehaus-memcache",
-    previousArtifact := youngestForwardCompatible("memcache"),
-    libraryDependencies += Finagle.module("memcached")
+  lazy val storehausMemcache = module("memcache").settings(
+    libraryDependencies ++= Seq(
+      "com.twitter" %% "algebird-core" % algebirdVersion,
+      "com.twitter" %% "bijection-core" % bijectionVersion,
+      "com.twitter" %% "bijection-netty" % bijectionVersion,
+      Finagle.module("memcached")
+    )
   ).dependsOn(storehausAlgebra % "test->test;compile->compile")
 
-  lazy val storehausMySQL = Project(
-    id = "storehaus-mysql",
-    base = file("storehaus-mysql"),
-    settings = sharedSettings
-  ).settings(
-    name := "storehaus-mysql",
-    previousArtifact := youngestForwardCompatible("mysql"),
-    libraryDependencies += Finagle.module("mysql", "6.2.1") // tests fail with the latest
-  ).dependsOn(storehausCore % "test->test;compile->compile")
+  lazy val storehausMySQL = module("mysql").settings(
+    libraryDependencies += Finagle.module("mysql")
+  ).dependsOn(storehausAlgebra % "test->test;compile->compile")
 
-  lazy val storehausRedis = Project(
-    id = "storehaus-redis",
-    base = file("storehaus-redis"),
-    settings = sharedSettings
-  ).settings(
-    name := "storehaus-redis",
-    previousArtifact := youngestForwardCompatible("redis"),
-    libraryDependencies += Finagle.module("redis"),
+  lazy val storehausRedis = module("redis").settings(
+    libraryDependencies ++= Seq (
+      "com.twitter" %% "bijection-core" % bijectionVersion,
+      "com.twitter" %% "bijection-netty" % bijectionVersion,
+      Finagle.module("redis")
+    ),
     // we don't want various tests clobbering each others keys
-    parallelExecution in Test := false,
-    testOptions in Test += Tests.Cleanup { loader =>
-      val c = loader.loadClass("com.twitter.storehaus.redis.Cleanup$")
-      c.getMethod("cleanup").invoke(c.getField("MODULE$").get(c))
-    }
+    parallelExecution in Test := false
   ).dependsOn(storehausAlgebra % "test->test;compile->compile")
 
   lazy val storehausZookeeper = Project(
@@ -175,9 +193,46 @@ object StorehausBuild extends Build {
     name := "storehaus-zookeeper",
     previousArtifact := youngestForwardCompatible("zookeeper"),
     libraryDependencies +=
-      ("com.twitter" %% "util-zk" % "6.3.0")
+      ("com.twitter" %% "util-zk" % utilVersion)
        .exclude("javax.jms", "jms")
        .exclude("com.sun.jdmk", "jmxtools")
        .exclude("com.sun.jmx", "jmxri")
   ).dependsOn(storehausCore % "test->test;compile->compile")
+
+  lazy val storehausHBase= module("hbase").settings(
+    libraryDependencies ++= Seq(
+      "com.twitter" %% "algebird-core" % algebirdVersion,
+      "com.twitter" %% "bijection-core" % bijectionVersion,
+      "com.twitter" %% "bijection-hbase" % bijectionVersion ,
+      "org.hbase" % "asynchbase" % "1.4.1" % "provided->default" intransitive(),
+      "com.stumbleupon" % "async" % "1.4.0" % "provided->default" intransitive(),
+      "org.apache.hbase" % "hbase" % "0.94.6" % "provided->default" classifier "tests" classifier "",
+      "org.apache.hadoop" % "hadoop-core" % "1.2.0" % "provided->default",
+      "org.apache.hadoop" % "hadoop-test" % "1.2.0" % "test"
+    ),
+    parallelExecution in Test := false
+  ).dependsOn(storehausAlgebra % "test->test;compile->compile")
+
+  lazy val storehausDynamoDB= module("dynamodb").settings(
+    libraryDependencies ++= Seq(
+      "com.twitter" %% "algebird-core" % algebirdVersion,
+      "com.twitter" %% "bijection-core" % bijectionVersion,
+      "com.twitter" %% "bijection-hbase" % bijectionVersion ,
+      "com.amazonaws" % "aws-java-sdk" % "1.5.7"
+      ////use alternator for local testing
+      //"com.michelboudreau" % "alternator" % "0.6.4" % "test"
+    ),
+    parallelExecution in Test := false
+  ).dependsOn(storehausAlgebra % "test->test;compile->compile")
+
+  val storehausTesting = Project(
+    id = "storehaus-testing",
+    base = file("storehaus-testing"),
+    settings = sharedSettings ++ Seq(
+      name := "storehaus-testing",
+      previousArtifact := youngestForwardCompatible("testing"),
+      libraryDependencies ++= Seq("org.scalacheck" %% "scalacheck" % "1.10.0" withSources(),
+        withCross("com.twitter" %% "util-core" % utilVersion))
+    )
+  )
 }