+++ storehaus-mysql/src/test/scala/com/twitter/storehaus/mysql/MySQLStoreProperties.scala
@@ -19,17 +19,17 @@ package com.twitter.storehaus.mysql
 import java.util.logging.Level
 
 import com.twitter.finagle.exp.mysql.Client
-import com.twitter.util.Await
+import com.twitter.storehaus.testing.SelfAggregatingCloseableCleanup
+import com.twitter.storehaus.testing.generator.NonEmpty
+import com.twitter.util.{Await, Future}
 
 import org.scalacheck.Arbitrary
 import org.scalacheck.Gen
 import org.scalacheck.Properties
 import org.scalacheck.Prop.forAll
 
-object MySqlStoreProperties extends Properties("MySqlStore") {
-
-  // used to generate arbitrary pairs of types we want to test
-  def validPairs[T: Arbitrary] = Arbitrary.arbitrary[List[(T, Option[T])]] suchThat(!_.isEmpty)
+object MySqlStoreProperties extends Properties("MySqlStore")
+  with SelfAggregatingCloseableCleanup[MySqlStore] {
 
   def put(s: MySqlStore, pairs: List[(MySqlValue, Option[MySqlValue])]) {
     pairs.foreach { case (k, v) =>
@@ -37,11 +37,20 @@ object MySqlStoreProperties extends Properties("MySqlStore") {
     }
   }
 
-  def putAndGetStoreTest(store: MySqlStore, pairs: Gen[List[(Any, Option[Any])]] = validPairs[String]) =
+  def multiPut(s: MySqlStore, pairs: List[(MySqlValue, Option[MySqlValue])]) {
+    Await.result(Future.collect(s.multiPut(pairs.toMap).values.toSeq))
+  }
+
+  /** invert any type to MySql String values. Because most mysql configuraions are case insensitive by default,
+   *  we lowercase key's here for normalization */
+  def stringify(examples: List[(Any, Option[Any])]) =
+    examples.map { case (k, v) =>
+      (MySqlStringInjection.invert(k.toString.toLowerCase).get, v.flatMap { d => MySqlStringInjection.invert(d.toString).toOption })
+    }
+
+  def putAndGetStoreTest(store: MySqlStore, pairs: Gen[List[(Any, Option[Any])]] = NonEmpty.Pairing.alphaStrs()) =
     forAll(pairs) { (examples: List[(Any, Option[Any])]) =>
-      val stringified = examples.map { case (k, v) =>
-          (MySqlStringInjection.invert(k.toString).get, v.flatMap { d => MySqlStringInjection.invert(d.toString) })
-        }
+      val stringified = stringify(examples)
       put(store, stringified)
       stringified.toMap.forall { case (k, optV) =>
         val foundOptV = Await.result(store.get(k))
@@ -49,12 +58,10 @@ object MySqlStoreProperties extends Properties("MySqlStore") {
       }
     }
 
-  def putAndMultiGetStoreTest(store: MySqlStore, pairs: Gen[List[(Any, Option[Any])]] = validPairs[String]) =
+  def multiPutAndMultiGetStoreTest(store: MySqlStore, pairs: Gen[List[(Any, Option[Any])]] = NonEmpty.Pairing.alphaStrs()) =
     forAll(pairs) { (examples: List[(Any, Option[Any])]) =>
-      val stringified = examples.map { case (k, v) =>
-          (MySqlStringInjection.invert(k.toString).get, v.map { case value => MySqlStringInjection.invert(value.toString).get })
-        }
-      put(store, stringified)
+      val stringified = stringify(examples)
+      multiPut(store, stringified)
       val data = stringified.toMap
       val result = store.multiGet(data.keySet)
       data.forall { case (k, optV) =>
@@ -69,7 +76,7 @@ object MySqlStoreProperties extends Properties("MySqlStore") {
 
   def compareValues(k: MySqlValue, expectedOptV: Option[MySqlValue], foundOptV: Option[MySqlValue]) = {
     val isMatch = expectedOptV match {
-      case Some(value) => !foundOptV.isEmpty && foundOptV.get == value 
+      case Some(value) => !foundOptV.isEmpty && foundOptV.get == value
       case None => foundOptV.isEmpty
     }
     if (!isMatch) printErr(k, expectedOptV, foundOptV)
@@ -92,32 +99,32 @@ object MySqlStoreProperties extends Properties("MySqlStore") {
     withStore(putAndGetStoreTest(_), "text", "blob")
 
   property("MySqlStore text->text multiget") =
-    withStore(putAndMultiGetStoreTest(_), "text", "text", true)
+    withStore(multiPutAndMultiGetStoreTest(_), "text", "text", true)
 
   property("MySqlStore blob->blob multiget") =
-    withStore(putAndMultiGetStoreTest(_), "blob", "blob", true)
+    withStore(multiPutAndMultiGetStoreTest(_), "blob", "blob", true)
 
   property("MySqlStore text->blob multiget") =
-    withStore(putAndMultiGetStoreTest(_), "text", "blob", true)
+    withStore(multiPutAndMultiGetStoreTest(_), "text", "blob", true)
 
   property("MySqlStore int->int") =
-    withStore(putAndGetStoreTest(_, validPairs[Int]), "int", "int")
+    withStore(putAndGetStoreTest(_, NonEmpty.Pairing.numerics[Int]()), "int", "int")
 
   property("MySqlStore int->int multiget") =
-    withStore(putAndMultiGetStoreTest(_, validPairs[Int]), "int", "int", true)
+    withStore(multiPutAndMultiGetStoreTest(_, NonEmpty.Pairing.numerics[Int]()), "int", "int", true)
 
   property("MySqlStore bigint->bigint") =
-    withStore(putAndGetStoreTest(_, validPairs[Long]), "bigint", "bigint")
+    withStore(putAndGetStoreTest(_, NonEmpty.Pairing.numerics[Long]()), "bigint", "bigint")
 
   property("MySqlStore bigint->bigint multiget") =
-    withStore(putAndMultiGetStoreTest(_, validPairs[Long]), "bigint", "bigint", true)
+    withStore(multiPutAndMultiGetStoreTest(_, NonEmpty.Pairing.numerics[Long]()), "bigint", "bigint", true)
 
   property("MySqlStore smallint->smallint") =
-    withStore(putAndGetStoreTest(_, validPairs[Short]), "smallint", "smallint")
+    withStore(putAndGetStoreTest(_, NonEmpty.Pairing.numerics[Short]()), "smallint", "smallint")
 
   property("MySqlStore smallint->smallint multiget") =
-    withStore(putAndMultiGetStoreTest(_, validPairs[Short]), "smallint", "smallint", true)
-  
+    withStore(multiPutAndMultiGetStoreTest(_, NonEmpty.Pairing.numerics[Short]()), "smallint", "smallint", true)
+
   private def withStore[T](f: MySqlStore => T, kColType: String, vColType: String, multiGet: Boolean = false): T = {
     val client = Client("localhost:3306", "storehaususer", "test1234", "storehaus_test", Level.WARNING)
     // these should match mysql setup used in .travis.yml
@@ -126,9 +133,9 @@ object MySqlStoreProperties extends Properties("MySqlStore") {
     val schema = "CREATE TEMPORARY TABLE IF NOT EXISTS `"+tableName+"` (`key` "+kColType+" DEFAULT NULL, `value` "+vColType+" DEFAULT NULL) ENGINE=InnoDB DEFAULT CHARSET=utf8;"
     Await.result(client.query(schema))
 
-    val store = MySqlStore(client, tableName, "key", "value")
-    val result = f(store)
-    store.close
-    result
+    f(newStore(client, tableName))
   }
+
+  def newStore(client: Client, tableName: String) =
+    aggregateCloseable(MySqlStore(client, tableName, "key", "value"))
 }