+++ storehaus-redis/src/test/scala/com/twitter/storehaus/redis/RedisStoreProperties.scala
@@ -17,27 +17,23 @@
 package com.twitter.storehaus.redis
 
 import com.twitter.bijection.Injection
-import com.twitter.finagle.redis.Client
 import com.twitter.finagle.redis.util.{ CBToString, StringToChannelBuffer }
 import com.twitter.storehaus.{ FutureOps, Store }
 import com.twitter.storehaus.algebra.ConvertedStore
+import com.twitter.storehaus.testing.CloseableCleanup
+import com.twitter.storehaus.testing.generator.NonEmpty
 import com.twitter.util.Await
 import org.jboss.netty.buffer.ChannelBuffer
 import org.scalacheck.{ Arbitrary, Gen, Properties }
 import org.scalacheck.Prop._
-import scala.util.control.Exception.allCatch
+import scala.util.Try
 
 object RedisStoreProperties extends Properties("RedisStore")
   with CloseableCleanup[Store[String, String]]
   with DefaultRedisClient {
 
-  def paired: Gen[(String, Option[String])] = for {
-    str <- Generators.nonEmptyAlphaStr
-    opt <- Generators.nonEmptyAlphaStrOpt
-  } yield (str, opt)
-
   def validPairs: Gen[List[(String, Option[String])]] =
-    Gen.listOfN(10,paired)
+    NonEmpty.Pairing.alphaStrs()
 
   def baseTest[K : Arbitrary, V: Arbitrary: Equiv](store: Store[K, V], validPairs: Gen[List[(K, Option[V])]])
     (put: (Store[K, V], List[(K, Option[V])]) => Unit) =
@@ -62,14 +58,12 @@ object RedisStoreProperties extends Properties("RedisStore")
   def storeTest(store: Store[String, String]) =
     putStoreTest(store, validPairs) && multiPutStoreTest(store, validPairs)
 
-  implicit def strToCb = new Injection[String, ChannelBuffer] {
-    def apply(a: String): ChannelBuffer = StringToChannelBuffer(a)
-    def invert(b: ChannelBuffer): Option[String] = allCatch.opt(CBToString(b))
-  }
+  implicit def strToCb =
+    Injection.build(StringToChannelBuffer(_: String))(
+      (b: ChannelBuffer) => Try(CBToString(b)))
+
   val closeable =
     RedisStore(client).convert(StringToChannelBuffer(_: String))
 
-
- property("RedisStore test") =
-   storeTest(closeable)
+  property("RedisStore test") = storeTest(closeable)
 }